update default values

+++ src/main/api/controllers/customers.ts
@@ -1,6 +1,8 @@
 import { Request, Response, NextFunction } from 'express';
 import Customer from '@api/models/customers';
 
+const DEFAULT_ID = '0a0aaa0a0aa00000aaaaaa0a';
+
 const getAllCustomers = async (_req: Request, res: Response, next: NextFunction) => {
   try {
     const customers = await Customer.find();
@@ -11,7 +13,7 @@ const getAllCustomers = async (_req: Request, res: Response, next: NextFunction)
 }
 const getAllClients = async (_req: Request, res: Response, next: NextFunction) => {
   try {
-    const customers = await Customer.find({ type: 'Client' });
+    const customers = await Customer.find({ type: 'Client', _id: { $ne: DEFAULT_ID } });
     return res.status(200).send(customers);
   } catch (error) {
     return next(error);
@@ -19,7 +21,7 @@ const getAllClients = async (_req: Request, res: Response, next: NextFunction) =
 }
 const getAllSuppliers = async (_req: Request, res: Response, next: NextFunction) => {
   try {
-    const customers = await Customer.find({ type: 'Supplier' });
+    const customers = await Customer.find({ type: 'Supplier', _id: { $ne: DEFAULT_ID } });
     return res.status(200).send(customers);
   } catch (error) {
     return next(error);
@@ -39,6 +41,9 @@ const createNewCustomer = async (req: Request, res: Response, next: NextFunction
 const updateOne = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const { params: { id }, body: payload } = req;
+    if (id === DEFAULT_ID) {
+      return res.status(403).send({ message: 'Cannot edit the default customer' });
+    }
     const updatedCustomer = await Customer.findByIdAndUpdate(id, payload, { new: true });
     return res.status(200).send(updatedCustomer);
   } catch (error) {
@@ -49,6 +54,9 @@ const updateOne = async (req: Request, res: Response, next: NextFunction) => {
 const deleteOne = async (req: Request, res: Response, next: NextFunction) => {
   try {
     const { params: { id } } = req;
+    if (id === DEFAULT_ID) {
+      return res.status(403).send({ message: 'Cannot delete the default customer' });
+    }
     await Customer.findByIdAndDelete(id);
     return res.status(200).send({ success: true });
   } catch (error) {
